+++ src/main/java/nbbang/com/nbbang/global/socket/StompWebSocketMessageBrokerConfig.java
@@ -1,7 +1,7 @@
 package nbbang.com.nbbang.global.socket;
 
 import lombok.RequiredArgsConstructor;
-import nbbang.com.nbbang.global.socket.interceptor.LoginHandShakeInterceptor;
+import nbbang.com.nbbang.global.socket.handshakeinterceptor.LoginHandShakeInterceptor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.simp.config.ChannelRegistration;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
