+++ src/main/java/nbbang/com/nbbang/global/socket/handshakeinterceptor/SocketIdUtilImpl.java
@@ -1,6 +1,5 @@
-package nbbang.com.nbbang.global.socket.interceptor;
+package nbbang.com.nbbang.global.socket.handshakeinterceptor;
 
-import nbbang.com.nbbang.global.socket.interceptor.SocketIdUtil;
 import org.springframework.http.server.ServerHttpRequest;
 import org.springframework.http.server.ServletServerHttpRequest;
 import org.springframework.messaging.Message;
